Okay. Maybe I haven't finished exactly, but maybe now?? Added request if entered password is correct, updated tests and fix a very stupid bug

+++ backend/__init__.py
@@ -3,12 +3,14 @@ import os
 
 # Import the framework
 from flask import Flask
+from flask_cors import CORS
 from flask_restful import Resource, Api, reqparse
 from .database import *
 from base64 import b64decode
 
 # Create an instance of Flask
 app = Flask(__name__)
+CORS(app)
 
 # Create the API
 api = Api(app)
@@ -199,6 +201,8 @@ class Groups2(Resource):
         except ValueError:
             return 400, 400
         values = database.get_group_name(group_id=group_id)
+        parser.add_argument('Authorization', location='headers')
+
         if values[0]:
             return values[2], values[1]
         else:
@@ -212,8 +216,12 @@ class Groups2(Resource):
             group_id = int(group_id)
         except ValueError:
             return 400, 400
-        values = database.delete_group(group_id=group_id, password=password)
-        return values[1], values[1]
+        if args['Authorization']:
+            values = database.check_group_pass(group_id=group_id, password=decode_password(args['Authorization']))
+            return values[1], values[1]
+        else:
+            values = database.delete_group(group_id=group_id, password=password)
+            return values[1], values[1]
 
     def put(self, group_id):
         parser.add_argument('name')
